requirements.txt
flask>=3.0.0
bcrypt>=4.1.0
cryptography>=42.0.0
dnspython>=2.5.0
python-whois>=0.9.4
rich>=13.7.0
waitress>=3.0.0
openai>=1.30.0
google-generativeai>=0.5.0
requests>=2.31.0
markdown>=3.4.0
bleach>=6.1.0
